Rename image_learner folder to imagelearner and update MetaFormer models
