--- tools/image_learner/test-data/mnist_subset.csv
@@ -1,121 +1,0 @@
-image_path,label,split
-training/0/5680.jpg,0,0
-training/0/5699.jpg,0,0
-training/0/5766.jpg,0,0
-training/0/5524.jpg,0,0
-training/0/5003.jpg,0,0
-training/0/5527.jpg,0,0
-training/0/5359.jpg,0,0
-training/0/5452.jpg,0,0
-training/0/5010.jpg,0,0
-training/0/5405.jpg,0,0
-training/1/6100.jpg,1,0
-training/1/6015.jpg,1,0
-training/1/5754.jpg,1,0
-training/1/6275.jpg,1,0
-training/1/6247.jpg,1,0
-training/1/6552.jpg,1,0
-training/1/6129.jpg,1,0
-training/1/6733.jpg,1,0
-training/1/6590.jpg,1,0
-training/1/6727.jpg,1,0
-training/2/5585.jpg,2,0
-training/2/5865.jpg,2,0
-training/2/4984.jpg,2,0
-training/2/4992.jpg,2,0
-training/2/5008.jpg,2,0
-training/2/5325.jpg,2,0
-training/2/5438.jpg,2,0
-training/2/5807.jpg,2,0
-training/2/5323.jpg,2,0
-training/2/5407.jpg,2,0
-training/3/5869.jpg,3,0
-training/3/5333.jpg,3,0
-training/3/5813.jpg,3,0
-training/3/6093.jpg,3,0
-training/3/5714.jpg,3,0
-training/3/5519.jpg,3,0
-training/3/5586.jpg,3,0
-training/3/5410.jpg,3,0
-training/3/5577.jpg,3,0
-training/3/5710.jpg,3,0
-training/4/5092.jpg,4,0
-training/4/5793.jpg,4,0
-training/4/5610.jpg,4,0
-training/4/5123.jpg,4,0
-training/4/5685.jpg,4,0
-training/4/4972.jpg,4,0
-training/4/4887.jpg,4,0
-training/4/5052.jpg,4,0
-training/4/5348.jpg,4,0
-training/4/5368.jpg,4,0
-training/5/5100.jpg,5,0
-training/5/4442.jpg,5,0
-training/5/4745.jpg,5,0
-training/5/4592.jpg,5,0
-training/5/4707.jpg,5,0
-training/5/5305.jpg,5,0
-training/5/4506.jpg,5,0
-training/5/5118.jpg,5,0
-training/5/4888.jpg,5,0
-training/5/5282.jpg,5,0
-training/6/5553.jpg,6,0
-training/6/5260.jpg,6,0
-training/6/5899.jpg,6,0
-training/6/5231.jpg,6,0
-training/6/5743.jpg,6,0
-training/6/5567.jpg,6,0
-training/6/5823.jpg,6,0
-training/6/5849.jpg,6,0
-training/6/5076.jpg,6,0
-training/6/5435.jpg,6,0
-training/7/6036.jpg,7,0
-training/7/5488.jpg,7,0
-training/7/5506.jpg,7,0
-training/7/6194.jpg,7,0
-training/7/5934.jpg,7,0
-training/7/5634.jpg,7,0
-training/7/5834.jpg,7,0
-training/7/5721.jpg,7,0
-training/7/6204.jpg,7,0
-training/7/5481.jpg,7,0
-training/8/5844.jpg,8,0
-training/8/5001.jpg,8,0
-training/8/5785.jpg,8,0
-training/8/5462.jpg,8,0
-training/8/4938.jpg,8,0
-training/8/4933.jpg,8,0
-training/8/5341.jpg,8,0
-training/8/5057.jpg,8,0
-training/8/4880.jpg,8,0
-training/8/5039.jpg,8,0
-training/9/5193.jpg,9,0
-training/9/5870.jpg,9,0
-training/9/5756.jpg,9,0
-training/9/5186.jpg,9,0
-training/9/5688.jpg,9,0
-training/9/5579.jpg,9,0
-training/9/5444.jpg,9,0
-training/9/5931.jpg,9,0
-training/9/5541.jpg,9,0
-training/9/5786.jpg,9,0
-test/0/833.jpg,0,2
-test/0/855.jpg,0,2
-test/1/1110.jpg,1,2
-test/1/969.jpg,1,2
-test/2/961.jpg,2,2
-test/2/971.jpg,2,2
-test/3/895.jpg,3,2
-test/3/1005.jpg,3,2
-test/4/940.jpg,4,2
-test/4/975.jpg,4,2
-test/5/780.jpg,5,2
-test/5/834.jpg,5,2
-test/6/932.jpg,6,2
-test/6/796.jpg,6,2
-test/7/835.jpg,7,2
-test/7/863.jpg,7,2
-test/8/899.jpg,8,2
-test/8/898.jpg,8,2
-test/9/1007.jpg,9,2
-test/9/954.jpg,9,2
