--- tools/image_learner/test-data/utkface_labels.csv
@@ -1,13 +1,0 @@
-image_path,label,split
-1_0_4_20161221195047839.jpg.chip.jpg,1,0
-4_0_0_20170110213253207.jpg.chip.jpg,4,0
-18_1_0_20170109212647587.jpg.chip.jpg,18,0
-24_1_2_20170116164628750.jpg.chip.jpg,24,0
-26_1_0_20170116164911648.jpg.chip.jpg,26,0
-28_1_0_20170109141748286.jpg.chip.jpg,28,0
-31_0_4_20170120133240958.jpg.chip.jpg,31,0
-35_0_0_20170104202556995.jpg.chip.jpg,35,0
-42_0_2_20170117130543345.jpg.chip.jpg,42,1
-52_0_1_20170117175021585.jpg.chip.jpg,52,1
-62_1_3_20170109132000815.jpg.chip.jpg,62,2
-116_1_2_20170112220255503.jpg.chip.jpg,116,2
