+++ tools/imagelearner/MetaFormer/metaformer_stacked_cnn.py
