+++ tools/imagelearner/test-data/image_metadata.csv
